+++ subcommands/createV2/plugins/handleUIElement.js
@@ -9,13 +9,17 @@ const {
   generateUiElementIndexJs,
   generateUiElementsReadme,
   generateUiElementIndexHtml,
-  generateUiElementsEsLintRc,
+  generateUiElementEsLintRc,
   generateUiElementBootstrapJs,
   generateUiElementPackageJson,
   generateUiElementsPrettierRc,
   generateUiElementsCommitlintRc,
   generateUiElementFederationExpose,
   generateUiElementFederationShared,
+  generateUiElementBabelRc,
+  generateUiElementAppTestJs,
+  generateUiElementJestConfig,
+  generateUiElementJestSetup,
 } = require('../../../templates/createTemplates/uiElement-templates')
 
 // eslint-disable-next-line no-unused-vars
@@ -55,9 +59,13 @@ class handleUIElement {
         const fedExposeString = generateUiElementFederationExpose(blockName)
         const fedSharedString = generateUiElementFederationShared(blockName)
         const packageJsonString = generateUiElementPackageJson(blockName)
-        const eslintrcString = generateUiElementsEsLintRc()
+        const eslintrcString = generateUiElementEsLintRc()
         const commitLintRcString = generateUiElementsCommitlintRc()
         const prettierrcString = generateUiElementsPrettierRc()
+        const babelRcString = generateUiElementBabelRc()
+        const jestConfigString = generateUiElementJestConfig()
+        const jestSetupString = generateUiElementJestSetup()
+        const appTestString = generateUiElementAppTestJs()
 
         mkdirSync(`${core.blockFolderPath}/public`)
         mkdirSync(`${core.blockFolderPath}/src/remote`, { recursive: true })
@@ -78,6 +86,10 @@ class handleUIElement {
         writeFileSync(`${core.blockFolderPath}/.eslintrc.json`, eslintrcString)
         writeFileSync(`${core.blockFolderPath}/.prettierrc.json`, prettierrcString)
         writeFileSync(`${core.blockFolderPath}/.commitlintrc.json`, commitLintRcString)
+        writeFileSync(`${core.blockFolderPath}/.babelrc`, babelRcString)
+        writeFileSync(`${core.blockFolderPath}/jest.config.js`, jestConfigString)
+        writeFileSync(`${core.blockFolderPath}/jest.setup.js`, jestSetupString)
+        writeFileSync(`${core.blockFolderPath}/App.test.js`, appTestString)
       }
     )
   }