+++ subcommands/startV2/plugins/handleJSViewStart/index.js
@@ -1,3 +1,4 @@
+const chalk = require('chalk')
 const { startJsProgram, handleReportLog } = require('./utils')
 const singleBuild = require('./singleBuild')
 // eslint-disable-next-line no-unused-vars
@@ -48,11 +49,7 @@ class HandleJSViewStart {
         this.elementsPort = this.elementsBlocks[0]?.availablePort
         this.containerPort = this.containerBlocks[0]?.availablePort
 
-        // Release port before server start
-        this.elementsBlocks[0]?.portKey?.abort()
-        this.containerBlocks[0]?.portKey?.abort()
-
-        await singleBuild({
+        const { error } = await singleBuild({
           core,
           blocks: {
             elementsBlocks: this.elementsBlocks,
@@ -66,6 +63,7 @@ class HandleJSViewStart {
           env: core.cmdOpts.environment,
         })
 
+        if (error) console.log(chalk.yellow(error))
         return
       }