+++ templates/createTemplates/uiElement-templates/generate-packageJson.js
@@ -8,8 +8,8 @@ const generateUiElementPackageJson = (name) => `{
   "scripts": {
     "start": "webpack-dev-server",
     "build": "webpack --mode production",
-    "prepare": "npx husky install",
-    "test": "npx jest",
+    "test": "jest",
+    "coverage": "jest --coverage",
     "lint:fix": "npx eslint *.js --fix",
     "format": "npx prettier ./**/*{.js,.json} --write",
     "pre-commit": "npx lint-staged"
@@ -39,14 +39,20 @@ const generateUiElementPackageJson = (name) => `{
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-import": "2.25.2",
     "eslint-plugin-react": "^7.32.2",
-    "husky":"7.0.0",
+    "eslint-plugin-jest": "^27.2.1",
+    "husky": "^7.0.4",
+    "jest": "^28.1.0",
+    "jest-environment-jsdom": "^28.1.0",
     "lint-staged":"11.2.3",
     "prettier":"2.4.1",
     "@commitlint/cli": "15.0.0",
     "@commitlint/config-conventional": "15.0.0",
     "@babel/core": "7.15.0",
     "@babel/eslint-parser": "^7.21.3",
+    "@babel/preset-env": "^7.21.4",
     "@babel/preset-react": "7.14.5",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
     "babel-loader": "8.2.2",
     "dotenv": "^16.0.0",
     "html-webpack-plugin": "5.3.2",
@@ -60,42 +66,7 @@ const generateUiElementPackageJson = (name) => `{
 }
 `
 
-const generateUiElementPackageJsonWithoutLint = (name) => `
-  {
-    "name": "${name}",
-    "main": "index.js",
-    "license": "ISC",
-    "type": "module",
-    "scripts": {
-      "start": "webpack-dev-server",
-      "build": "webpack --mode production",
-      "serve": "webpack-cli serve"
-    },
-    "dependencies": {
-      "@appblocks/js-sdk": "^0.0.11",
-      "react": "^18.2.0",
-      "react-dom": "^18.2.0",
-      "react-error-boundary": "^3.1.4",
-      "react-query": "^3.39.2",
-      "react-redux": "^7.2.5",
-      "react-router-dom": "^6.9.0",
-      "redux": "^4.1.2",
-      "state-pool": "^0.8.1"
-    },
-    "devDependencies": {
-      "@babel/core": "7.15.0",
-      "@babel/preset-react": "7.14.5",
-      "babel-loader": "8.2.2",
-      "dotenv": "^16.0.0",
-      "html-webpack-plugin": "5.3.2",
-      "serve": "12.0.0",
-      "url-loader": "^4.1.1",
-      "webpack": "5.52.0",
-      "webpack-cli": "4.10.0",
-      "webpack-dev-server": "4.1.0",
-      "@appblocks/node-sdk": "0.0.4"
-    }
-  }
-    `
+module.exports = { generateUiElementPackageJson }
 
-module.exports = { generateUiElementPackageJson, generateUiElementPackageJsonWithoutLint }
+// removed husky
+// "prepare": "npx husky install",